Refactoring, moving some functions around
