+++ src/actions/buildindustry.ts
@@ -1,15 +1,12 @@
 import * as bge from "bge-core";
 
 import { Game } from "../game";
-import { Card } from "../objects/card";
 import { IResourceSources } from "../objects/gameboard";
 import { IndustryLocation } from "../objects/industrylocation";
-import { IndustryTile } from "../objects/industrytile";
-import { ResourceMarket } from "../objects/resourcemarket";
 import { ResourceToken } from "../objects/resourcetoken";
-import { LinkLocation } from "../objects/linklocation";
 import { Player } from "../player";
-import { Industry, Resource, Era, City, OVERBUILDABLE_INDUSTRIES } from "../types";
+import { Industry, Resource, Era, OVERBUILDABLE_INDUSTRIES } from "../types";
+import { consumeResources } from ".";
 
 const console = bge.Logger.get("build-industry");
 
@@ -200,47 +197,4 @@ function getBuildableIndustriesAtLocation(location: IndustryLocation, player: Pl
 	}
 
 	return result;
-}
-
-/**
- * Prompts the given player to select which resources to consume, up to the given amount.
- */
-export async function consumeResources(player: Player, destination: IndustryLocation | LinkLocation,
-	resource: Resource, amount: number, sources: IResourceSources, market?: ResourceMarket) {
-
-	while (sources.tiles.length > 0 && amount > 0) {
-		const distance = sources.tiles[0].distance;
-		const choices = new Set(sources.tiles.filter(x => x.distance === distance).map(x => x.tile));
-
-		let tile = await player.prompt.clickAny(choices, {
-			message: `Select ${(resource === Resource.Iron ? "an" : "a")} ${Resource[resource]} to consume`,
-			autoResolveIfSingle: true
-		});
-
-		sources.tiles.splice(sources.tiles.findIndex(x => x.tile === tile), 1);
-
-		console.info(`Consuming ${Resource[resource]} from ${tile.name}`);
-
-		await tile.consumeResource(destination.spentResources);
-
-		--amount;
-
-		await player.game.delay.beat();
-	}
-
-	if (amount > 0) {
-		if (market == null) {
-			throw new Error("No market was given!");
-		}
-
-		const cost = market.getCost(amount);
-
-		destination.spentResources.push(...market.takeRange(amount));
-
-		console.info(`Spending £${cost} to buy ${amount} ${Resource[resource]} from the market`);
-
-		player.spendMoney(cost);
-
-		await player.game.delay.beat();
-	}
 }
